Add semitranslucent backdrop for posters
